UPD | comments #3

+++ include/http/ManapiBaseHttp.hpp
@@ -35,7 +35,7 @@ namespace manapi::net::http::internal {
     void send_response_sync_cb (uq_handle_data_t cdata, std::unique_ptr<response> res, response_features_t features);
     void send_response_stream_cb (uq_handle_data_t cdata, std::unique_ptr<response> res, response_features_t features);
     void send_response_async_cb (uq_handle_data_t cdata, std::unique_ptr<response> res, response_features_t features);
-    future<ssize_t> mask_response (handle_data_t* cdata, response *res, bool finish);
+    future<int> mask_response (handle_data_t* cdata, response *res, bool finish);
     void handle_income_request (uq_handle_data_t cdata, int status);
     void send_error_response (uq_handle_data_t cdata, int status = http::INTERNAL_SERVER_ERROR_500);
     //future<void> send_file(uq_handle_data_t cdata, filesystem::fstream f, ssize_t size, std::vector<replace_founded_item> replacers);
